fix: three things a real sign in found, and the legal pages link where they should
Driving the console by hand, signed in as a fresh account, turned up:

The billing page died on a blank screen when the API is one deploy behind the
bundle, because a meter read `.toLocaleString()` off a field the older API
does not send. A missing figure is now a missing meter, not a missing page.
That shape happens on every rolling deploy and on any cached bundle.

Storage read "0 MB" for a workspace that had just built an app, because a few
hundred kilobytes rounded down to nothing. It carries one decimal now.

The register page promised terms and a privacy policy in plain text with
nothing to click. Both are real pages now, so both are links.

Also: "Basivo operates Basivo" on the terms page when the legal name and the
brand are the same word, and a test fixture that read a warmed sqlite file as
UTF-8, which only failed once this machine had a real warm cache.

Verified end to end on the way through: upload an image, send one message,
watch the free agent build it, and the photograph is in the built page, in the
preview, at the public address after Deploy, and gone from it after Unpublish.

+++ apps/api/basivo_orch/billing/schemas.py
@@ -51,8 +51,10 @@ class UsageRead(BaseModel):
     seats_used: int
     seats_limit: int | None
     #: Files and app builds, in megabytes, because a workspace reads its own
-    #: storage in the units its plan is written in.
-    storage_used_mb: int
+    #: storage in the units its plan is written in. One decimal: a workspace
+    #: holding a few hundred kilobytes has not used "0 MB", and a meter that
+    #: says it has looks broken to the person who just built something.
+    storage_used_mb: float
     storage_limit_mb: int | None
     history_days: int
